refactor: App 클래스 로직 LadderGame 클래스로 이동
